Some adjustments had been made to have more readable errors on the api side

+++ Source/Net.Clinic.Application/Features/Appointments/Delete/DeleteAppointmentCommandHandler.cs
@@ -4,7 +4,7 @@ using Net.Clinic.Domain.IRepositories;
 
 namespace Net.Clinic.Application.Features.Appointments.Delete
 {
-    public class DeleteAppointmentCommandHandler:IRequestHandler<DeleteAppointmentCommand, ApplicationResult<Unit>>
+    public class DeleteAppointmentCommandHandler : IRequestHandler<DeleteAppointmentCommand, ApplicationResult<Unit>>
     {
         private readonly IAppointmentRepository _repository;
 
@@ -12,13 +12,23 @@ namespace Net.Clinic.Application.Features.Appointments.Delete
         {
             this._repository = repository;
         }
-        
+
         public async Task<ApplicationResult<Unit>> Handle(DeleteAppointmentCommand request, CancellationToken cancellationToken)
         {
             var appointment = await this._repository.FetchAppointmentByIdAsync(request.AppointmentId, cancellationToken);
-            this._repository.RemoveAppointment(appointment);
-            await this._repository.SaveChangesAsync(cancellationToken);
-            return ApplicationResult<Unit>.Ok(Unit.Value);
+            var errors = new Dictionary<string, ICollection<string>>();
+
+            if (appointment == null)
+            {
+                errors.Add("AppointmentDelete", [$"Appointment {request.AppointmentId} not found"]);
+                return ApplicationResult<Unit>.Fail(errors);
+            }
+            else
+            {
+                this._repository.RemoveAppointment(appointment);
+                await this._repository.SaveChangesAsync(cancellationToken);
+                return ApplicationResult<Unit>.Ok(Unit.Value);
+            }
         }
     }
 }
